+++ com.unity.netcode.adapter.utp/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this package will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+- Rename the 'Send Queue Batch Size' property to 'Max Payload Size' to better reflect its usage.
+
+### Fixed
+
 ## [1.0.0-pre.4] - 2022-01-04
 
 ### Added